Gradient of the modified logLikelihood with common HPs for GPs in Magma

Gradient of the modified logLikelihood with common HPs for GPs in Magma

Usage

gr_GP_mod_common_hp(hp, db, mean, kern, post_cov, pen_diag)

Arguments

hp

A tibble or data frame containing hyper-parameters for all individuals.

db

A tibble containing the values we want to compute the logL on. Required columns: ID, Input, Output. Additional covariate columns are allowed.

mean

A vector, specifying the mean of the GPs at the reference inputs.

kern

A kernel function.

post_cov

A matrix, covariance parameter of the hyper-posterior. Used to compute the correction term.

pen_diag

A jitter term that is added to the covariance matrix to avoid numerical issues when inverting, in cases of nearly singular matrices.

Value

A named vector, corresponding to the value of the hyper-parameters' gradients for the modified Gaussian log-Likelihood involved in Magma with the 'common HP' setting.

Examples

db <- simu_db(N = 10, common_input = TRUE)
mean <- tibble::tibble(Input = unique(db$Input), Output = 0)
hp <- tibble::tibble(se_variance = 1, se_lengthscale = 0.5)
post_cov <- kern_to_cov(unique(db$Input), "SE", hp)
MagmaClustR:::gr_GP_mod_common_hp(hp, db, mean, "SE", post_cov, 0.001)
#>    se_variance se_lengthscale 
#>      -4667.382      19337.504