Compute a covariance matrix for multiple individuals

Compute the covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_cov(data, kern, hp, deriv = NULL)

Arguments

data

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), hp())
MagmaClustR:::list_kern_to_cov(db, "SE", hp)
#> $`1`
#>            1.7      3.65      3.85       5.8      8.05       8.3       8.6
#> 1.7  1.3896264 1.1572333 1.1124535 0.6187961 0.1995750 0.1707805 0.1405362
#> 3.65 1.1572333 1.3896264 1.3869538 1.1124535 0.5473300 0.4908628 0.4273310
#> 3.85 1.1124535 1.3869538 1.3896264 1.1572333 0.5945652 0.5357972 0.4691513
#> 5.8  0.6187961 1.1124535 1.1572333 1.3896264 1.0891440 1.0286439 0.9528664
#> 8.05 0.1995750 0.5473300 0.5945652 1.0891440 1.3896264 1.3854527 1.3695422
#> 8.3  0.1707805 0.4908628 0.5357972 1.0286439 1.3854527 1.3896264 1.3836203
#> 8.6  0.1405362 0.4273310 0.4691513 0.9528664 1.3695422 1.3836203 1.3896264
#> 8.85 0.1186786 0.3782052 0.4172209 0.8881041 1.3474766 1.3695422 1.3854527
#> 9    0.1069224 0.3504703 0.3877429 0.8489243 1.3305605 1.3572393 1.3789669
#> 9.35 0.0831195 0.2909481 0.3240667 0.7576867 1.2810760 1.3178148 1.3525118
#>           8.85         9      9.35
#> 1.7  0.1186786 0.1069224 0.0831195
#> 3.65 0.3782052 0.3504703 0.2909481
#> 3.85 0.4172209 0.3877429 0.3240667
#> 5.8  0.8881041 0.8489243 0.7576867
#> 8.05 1.3474766 1.3305605 1.2810760
#> 8.3  1.3695422 1.3572393 1.3178148
#> 8.6  1.3854527 1.3789669 1.3525118
#> 8.85 1.3896264 1.3881224 1.3730069
#> 9    1.3881224 1.3896264 1.3814578
#> 9.35 1.3730069 1.3814578 1.3896264
#> 
#> $`2`
#>            1.7      3.65      3.85       5.8      8.05       8.3       8.6
#> 1.7  1.3896264 1.1572333 1.1124535 0.6187961 0.1995750 0.1707805 0.1405362
#> 3.65 1.1572333 1.3896264 1.3869538 1.1124535 0.5473300 0.4908628 0.4273310
#> 3.85 1.1124535 1.3869538 1.3896264 1.1572333 0.5945652 0.5357972 0.4691513
#> 5.8  0.6187961 1.1124535 1.1572333 1.3896264 1.0891440 1.0286439 0.9528664
#> 8.05 0.1995750 0.5473300 0.5945652 1.0891440 1.3896264 1.3854527 1.3695422
#> 8.3  0.1707805 0.4908628 0.5357972 1.0286439 1.3854527 1.3896264 1.3836203
#> 8.6  0.1405362 0.4273310 0.4691513 0.9528664 1.3695422 1.3836203 1.3896264
#> 8.85 0.1186786 0.3782052 0.4172209 0.8881041 1.3474766 1.3695422 1.3854527
#> 9    0.1069224 0.3504703 0.3877429 0.8489243 1.3305605 1.3572393 1.3789669
#> 9.35 0.0831195 0.2909481 0.3240667 0.7576867 1.2810760 1.3178148 1.3525118
#>           8.85         9      9.35
#> 1.7  0.1186786 0.1069224 0.0831195
#> 3.65 0.3782052 0.3504703 0.2909481
#> 3.85 0.4172209 0.3877429 0.3240667
#> 5.8  0.8881041 0.8489243 0.7576867
#> 8.05 1.3474766 1.3305605 1.2810760
#> 8.3  1.3695422 1.3572393 1.3178148
#> 8.6  1.3854527 1.3789669 1.3525118
#> 8.85 1.3896264 1.3881224 1.3730069
#> 9    1.3881224 1.3896264 1.3814578
#> 9.35 1.3730069 1.3814578 1.3896264
#> 
#> $`3`
#>            1.7      3.65      3.85       5.8      8.05       8.3       8.6
#> 1.7  1.3896264 1.1572333 1.1124535 0.6187961 0.1995750 0.1707805 0.1405362
#> 3.65 1.1572333 1.3896264 1.3869538 1.1124535 0.5473300 0.4908628 0.4273310
#> 3.85 1.1124535 1.3869538 1.3896264 1.1572333 0.5945652 0.5357972 0.4691513
#> 5.8  0.6187961 1.1124535 1.1572333 1.3896264 1.0891440 1.0286439 0.9528664
#> 8.05 0.1995750 0.5473300 0.5945652 1.0891440 1.3896264 1.3854527 1.3695422
#> 8.3  0.1707805 0.4908628 0.5357972 1.0286439 1.3854527 1.3896264 1.3836203
#> 8.6  0.1405362 0.4273310 0.4691513 0.9528664 1.3695422 1.3836203 1.3896264
#> 8.85 0.1186786 0.3782052 0.4172209 0.8881041 1.3474766 1.3695422 1.3854527
#> 9    0.1069224 0.3504703 0.3877429 0.8489243 1.3305605 1.3572393 1.3789669
#> 9.35 0.0831195 0.2909481 0.3240667 0.7576867 1.2810760 1.3178148 1.3525118
#>           8.85         9      9.35
#> 1.7  0.1186786 0.1069224 0.0831195
#> 3.65 0.3782052 0.3504703 0.2909481
#> 3.85 0.4172209 0.3877429 0.3240667
#> 5.8  0.8881041 0.8489243 0.7576867
#> 8.05 1.3474766 1.3305605 1.2810760
#> 8.3  1.3695422 1.3572393 1.3178148
#> 8.6  1.3854527 1.3789669 1.3525118
#> 8.85 1.3896264 1.3881224 1.3730069
#> 9    1.3881224 1.3896264 1.3814578
#> 9.35 1.3730069 1.3814578 1.3896264
#>