Compute an inverse covariance matrix for multiple individuals

Compute the inverse covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_inv(db, kern, hp, pen_diag, deriv = NULL)

Arguments

db

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

pen_diag

A number. A jitter term, added on the diagonal to prevent numerical issues when inverting nearly singular matrices.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), MagmaClustR:::hp())
MagmaClustR:::list_kern_to_inv(db, "SE", hp, 0)
#> $`1`
#>               1.2           1.8          2.25          2.45           2.9
#> 1.2    1902991571  -19944440578  9.010025e+10 -9.043424e+10   18852124630
#> 1.8  -19944440578  209526574833 -9.481984e+11  9.524389e+11 -198883288163
#> 2.25  90100251336 -948198366597  4.296508e+12 -4.318151e+12  902822097179
#> 2.45 -90434237724  952438850251 -4.318151e+12  4.340978e+12 -908094994703
#> 2.9   18852124630 -198883288163  9.028221e+11 -9.080950e+11  190199330512
#> 5.6   -1233539464   13136651458 -6.005608e+10  6.059669e+10  -12781745402
#> 7.15   4191112070  -44852276048  2.058132e+11 -2.080127e+11   44042952692
#> 7.7   -8744088757   93731578940 -4.306489e+11  4.354991e+11  -92328788681
#> 7.95   5454399656  -58511009488  2.689802e+11 -2.720788e+11   57716120699
#> 9.05   -144566909    1555675165 -7.168853e+09  7.259355e+09   -1543778852
#>               5.6          7.15           7.7          7.95        9.05
#> 1.2   -1233539464    4191112070   -8744088757    5454399656  -144566909
#> 1.8   13136651458  -44852276048   93731578940  -58511009488  1555675165
#> 2.25 -60056078466  205813169322 -430648931584  268980162035 -7168853095
#> 2.45  60596690621 -208012740844  435499117767 -272078832136  7259355245
#> 2.9  -12781745402   44042952692  -92328788681   57716120699 -1543778852
#> 5.6     895775865   -3160493213    6680550573   -4191774377   113963188
#> 7.15  -3160493213   11310606127  -24031306710   15114205009  -415237087
#> 7.7    6680550573  -24031306710   51155362997  -32201634096   888162696
#> 7.95  -4191774377   15114205009  -32201634096   20278678401  -560331654
#> 9.05    113963188    -415237087     888162696    -560331654    15611926
#> 
#> $`2`
#>               1.2           1.8          2.25          2.45           2.9
#> 1.2    1902991571  -19944440578  9.010025e+10 -9.043424e+10   18852124630
#> 1.8  -19944440578  209526574833 -9.481984e+11  9.524389e+11 -198883288163
#> 2.25  90100251336 -948198366597  4.296508e+12 -4.318151e+12  902822097179
#> 2.45 -90434237724  952438850251 -4.318151e+12  4.340978e+12 -908094994703
#> 2.9   18852124630 -198883288163  9.028221e+11 -9.080950e+11  190199330512
#> 5.6   -1233539464   13136651458 -6.005608e+10  6.059669e+10  -12781745402
#> 7.15   4191112070  -44852276048  2.058132e+11 -2.080127e+11   44042952692
#> 7.7   -8744088757   93731578940 -4.306489e+11  4.354991e+11  -92328788681
#> 7.95   5454399656  -58511009488  2.689802e+11 -2.720788e+11   57716120699
#> 9.05   -144566909    1555675165 -7.168853e+09  7.259355e+09   -1543778852
#>               5.6          7.15           7.7          7.95        9.05
#> 1.2   -1233539464    4191112070   -8744088757    5454399656  -144566909
#> 1.8   13136651458  -44852276048   93731578940  -58511009488  1555675165
#> 2.25 -60056078466  205813169322 -430648931584  268980162035 -7168853095
#> 2.45  60596690621 -208012740844  435499117767 -272078832136  7259355245
#> 2.9  -12781745402   44042952692  -92328788681   57716120699 -1543778852
#> 5.6     895775865   -3160493213    6680550573   -4191774377   113963188
#> 7.15  -3160493213   11310606127  -24031306710   15114205009  -415237087
#> 7.7    6680550573  -24031306710   51155362997  -32201634096   888162696
#> 7.95  -4191774377   15114205009  -32201634096   20278678401  -560331654
#> 9.05    113963188    -415237087     888162696    -560331654    15611926
#> 
#> $`3`
#>               1.2           1.8          2.25          2.45           2.9
#> 1.2    1902991571  -19944440578  9.010025e+10 -9.043424e+10   18852124630
#> 1.8  -19944440578  209526574833 -9.481984e+11  9.524389e+11 -198883288163
#> 2.25  90100251336 -948198366597  4.296508e+12 -4.318151e+12  902822097179
#> 2.45 -90434237724  952438850251 -4.318151e+12  4.340978e+12 -908094994703
#> 2.9   18852124630 -198883288163  9.028221e+11 -9.080950e+11  190199330512
#> 5.6   -1233539464   13136651458 -6.005608e+10  6.059669e+10  -12781745402
#> 7.15   4191112070  -44852276048  2.058132e+11 -2.080127e+11   44042952692
#> 7.7   -8744088757   93731578940 -4.306489e+11  4.354991e+11  -92328788681
#> 7.95   5454399656  -58511009488  2.689802e+11 -2.720788e+11   57716120699
#> 9.05   -144566909    1555675165 -7.168853e+09  7.259355e+09   -1543778852
#>               5.6          7.15           7.7          7.95        9.05
#> 1.2   -1233539464    4191112070   -8744088757    5454399656  -144566909
#> 1.8   13136651458  -44852276048   93731578940  -58511009488  1555675165
#> 2.25 -60056078466  205813169322 -430648931584  268980162035 -7168853095
#> 2.45  60596690621 -208012740844  435499117767 -272078832136  7259355245
#> 2.9  -12781745402   44042952692  -92328788681   57716120699 -1543778852
#> 5.6     895775865   -3160493213    6680550573   -4191774377   113963188
#> 7.15  -3160493213   11310606127  -24031306710   15114205009  -415237087
#> 7.7    6680550573  -24031306710   51155362997  -32201634096   888162696
#> 7.95  -4191774377   15114205009  -32201634096   20278678401  -560331654
#> 9.05    113963188    -415237087     888162696    -560331654    15611926
#>