Simulate a batch a data

Simulate a batch a output data, corresponding to one individual, coming from a GP with a the Squared Exponential kernel as covariance structure, and specified hyper-parameters and input.

Usage

simu_indiv_se(ID, input, covariate, mean, v, l, sigma)

Arguments

ID

An identification code, whether numeric or character.

input

A vector of numbers. The input variable that is used as 'reference' for input and outputs.

covariate

A vector of numbers. An additional input variable, observed along with each reference input.

mean

A vector of numbers. Prior mean values of the GP.

v

A number. The variance hyper-parameter of the SE kernel.

l

A number. The lengthscale hyper-parameter of the SE kernel.

sigma

A number. The noise hyper-parameter.

Value

A tibble containing a batch of output data along with input and additional information for a simulated individual.

Examples

MagmaClustR:::simu_indiv_se("A", 1:10, 0, rep(0, 10), 2, 1, 0.5)
#> # A tibble: 10 × 7
#>    ID     Output Input Covariate se_variance se_lengthscale noise
#>    <chr>   <dbl> <int>     <dbl>       <dbl>          <dbl> <dbl>
#>  1 A      2.25       1         0           2              1   0.5
#>  2 A      0.391      2         0           2              1   0.5
#>  3 A      1.26       3         0           2              1   0.5
#>  4 A      2.88       4         0           2              1   0.5
#>  5 A      0.0561     5         0           2              1   0.5
#>  6 A     -0.0720     6         0           2              1   0.5
#>  7 A     -0.226      7         0           2              1   0.5
#>  8 A     -2.61       8         0           2              1   0.5
#>  9 A     -0.279      9         0           2              1   0.5
#> 10 A     -3.22      10         0           2              1   0.5
MagmaClustR:::simu_indiv_se("B", 1:10, 2:11, 3:12, 1, 1, 1)
#> # A tibble: 10 × 7
#>    ID    Output Input Covariate se_variance se_lengthscale noise
#>    <chr>  <dbl> <int>     <int>       <dbl>          <dbl> <dbl>
#>  1 B       1.54     1         2           1              1     1
#>  2 B       5.05     2         3           1              1     1
#>  3 B       6.37     3         4           1              1     1
#>  4 B       9.67     4         5           1              1     1
#>  5 B      10.4      5         6           1              1     1
#>  6 B      13.5      6         7           1              1     1
#>  7 B      17.0      7         8           1              1     1
#>  8 B      21.4      8         9           1              1     1
#>  9 B      24.7      9        10           1              1     1
#> 10 B      24.2     10        11           1              1     1
MagmaClustR:::simu_indiv_se("C", 1:10, 5, rep(0, 10), 2, 1, 0.5)
#> # A tibble: 10 × 7
#>    ID    Output Input Covariate se_variance se_lengthscale noise
#>    <chr>  <dbl> <int>     <dbl>       <dbl>          <dbl> <dbl>
#>  1 C      6.99      1         5           2              1   0.5
#>  2 C      3.88      2         5           2              1   0.5
#>  3 C      3.64      3         5           2              1   0.5
#>  4 C      0.318     4         5           2              1   0.5
#>  5 C      2.78      5         5           2              1   0.5
#>  6 C      3.82      6         5           2              1   0.5
#>  7 C      6.22      7         5           2              1   0.5
#>  8 C      8.81      8         5           2              1   0.5
#>  9 C      9.46      9         5           2              1   0.5
#> 10 C      5.85     10         5           2              1   0.5