Compute a covariance matrix for multiple individuals

Compute the covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_cov(data, kern, hp, deriv = NULL)

Arguments

data

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), hp())
MagmaClustR:::list_kern_to_cov(db, "SE", hp)
#> $`1`
#>              0.05         0.45          0.9          3.4          5.7
#> 0.05 1.331152e+01 1.267651e+01 1.067508e+01 4.318154e-01 7.743806e-04
#> 0.45 1.267651e+01 1.331152e+01 1.251299e+01 9.324715e-01 2.933671e-03
#> 0.9  1.067508e+01 1.251299e+01 1.331152e+01 1.972503e+00 1.167959e-02
#> 3.4  4.318154e-01 9.324715e-01 1.972503e+00 1.331152e+01 2.644748e+00
#> 5.7  7.743806e-04 2.933671e-03 1.167959e-02 2.644748e+00 1.331152e+01
#> 7.25 1.763713e-06 9.758884e-06 5.949716e-05 1.437716e-01 6.389683e+00
#> 7.65 2.890682e-07 1.763713e-06 1.200291e-05 5.343313e-02 4.166163e+00
#> 8.8  9.255605e-10 7.479839e-09 6.983417e-08 1.800799e-03 7.066987e-01
#> 9.4  3.354049e-11 3.138631e-10 3.455888e-09 2.228322e-04 2.032011e-01
#> 9.7  5.879358e-12 5.920282e-11 7.079192e-10 7.217923e-05 1.003344e-01
#>              7.25         7.65          8.8          9.4          9.7
#> 0.05 1.763713e-06 2.890682e-07 9.255605e-10 3.354049e-11 5.879358e-12
#> 0.45 9.758884e-06 1.763713e-06 7.479839e-09 3.138631e-10 5.920282e-11
#> 0.9  5.949716e-05 1.200291e-05 6.983417e-08 3.455888e-09 7.079192e-10
#> 3.4  1.437716e-01 5.343313e-02 1.800799e-03 2.228322e-04 7.217923e-05
#> 5.7  6.389683e+00 4.166163e+00 7.066987e-01 2.032011e-01 1.003344e-01
#> 7.25 1.331152e+01 1.267651e+01 6.389683e+00 3.242977e+00 2.127425e+00
#> 7.65 1.267651e+01 1.331152e+01 8.887237e+00 5.222936e+00 3.686946e+00
#> 8.8  6.389683e+00 8.887237e+00 1.331152e+01 1.192518e+01 1.039350e+01
#> 9.4  3.242977e+00 5.222936e+00 1.192518e+01 1.331152e+01 1.295051e+01
#> 9.7  2.127425e+00 3.686946e+00 1.039350e+01 1.295051e+01 1.331152e+01
#> 
#> $`2`
#>              0.05         0.45          0.9          3.4          5.7
#> 0.05 1.331152e+01 1.267651e+01 1.067508e+01 4.318154e-01 7.743806e-04
#> 0.45 1.267651e+01 1.331152e+01 1.251299e+01 9.324715e-01 2.933671e-03
#> 0.9  1.067508e+01 1.251299e+01 1.331152e+01 1.972503e+00 1.167959e-02
#> 3.4  4.318154e-01 9.324715e-01 1.972503e+00 1.331152e+01 2.644748e+00
#> 5.7  7.743806e-04 2.933671e-03 1.167959e-02 2.644748e+00 1.331152e+01
#> 7.25 1.763713e-06 9.758884e-06 5.949716e-05 1.437716e-01 6.389683e+00
#> 7.65 2.890682e-07 1.763713e-06 1.200291e-05 5.343313e-02 4.166163e+00
#> 8.8  9.255605e-10 7.479839e-09 6.983417e-08 1.800799e-03 7.066987e-01
#> 9.4  3.354049e-11 3.138631e-10 3.455888e-09 2.228322e-04 2.032011e-01
#> 9.7  5.879358e-12 5.920282e-11 7.079192e-10 7.217923e-05 1.003344e-01
#>              7.25         7.65          8.8          9.4          9.7
#> 0.05 1.763713e-06 2.890682e-07 9.255605e-10 3.354049e-11 5.879358e-12
#> 0.45 9.758884e-06 1.763713e-06 7.479839e-09 3.138631e-10 5.920282e-11
#> 0.9  5.949716e-05 1.200291e-05 6.983417e-08 3.455888e-09 7.079192e-10
#> 3.4  1.437716e-01 5.343313e-02 1.800799e-03 2.228322e-04 7.217923e-05
#> 5.7  6.389683e+00 4.166163e+00 7.066987e-01 2.032011e-01 1.003344e-01
#> 7.25 1.331152e+01 1.267651e+01 6.389683e+00 3.242977e+00 2.127425e+00
#> 7.65 1.267651e+01 1.331152e+01 8.887237e+00 5.222936e+00 3.686946e+00
#> 8.8  6.389683e+00 8.887237e+00 1.331152e+01 1.192518e+01 1.039350e+01
#> 9.4  3.242977e+00 5.222936e+00 1.192518e+01 1.331152e+01 1.295051e+01
#> 9.7  2.127425e+00 3.686946e+00 1.039350e+01 1.295051e+01 1.331152e+01
#> 
#> $`3`
#>              0.05         0.45          0.9          3.4          5.7
#> 0.05 1.331152e+01 1.267651e+01 1.067508e+01 4.318154e-01 7.743806e-04
#> 0.45 1.267651e+01 1.331152e+01 1.251299e+01 9.324715e-01 2.933671e-03
#> 0.9  1.067508e+01 1.251299e+01 1.331152e+01 1.972503e+00 1.167959e-02
#> 3.4  4.318154e-01 9.324715e-01 1.972503e+00 1.331152e+01 2.644748e+00
#> 5.7  7.743806e-04 2.933671e-03 1.167959e-02 2.644748e+00 1.331152e+01
#> 7.25 1.763713e-06 9.758884e-06 5.949716e-05 1.437716e-01 6.389683e+00
#> 7.65 2.890682e-07 1.763713e-06 1.200291e-05 5.343313e-02 4.166163e+00
#> 8.8  9.255605e-10 7.479839e-09 6.983417e-08 1.800799e-03 7.066987e-01
#> 9.4  3.354049e-11 3.138631e-10 3.455888e-09 2.228322e-04 2.032011e-01
#> 9.7  5.879358e-12 5.920282e-11 7.079192e-10 7.217923e-05 1.003344e-01
#>              7.25         7.65          8.8          9.4          9.7
#> 0.05 1.763713e-06 2.890682e-07 9.255605e-10 3.354049e-11 5.879358e-12
#> 0.45 9.758884e-06 1.763713e-06 7.479839e-09 3.138631e-10 5.920282e-11
#> 0.9  5.949716e-05 1.200291e-05 6.983417e-08 3.455888e-09 7.079192e-10
#> 3.4  1.437716e-01 5.343313e-02 1.800799e-03 2.228322e-04 7.217923e-05
#> 5.7  6.389683e+00 4.166163e+00 7.066987e-01 2.032011e-01 1.003344e-01
#> 7.25 1.331152e+01 1.267651e+01 6.389683e+00 3.242977e+00 2.127425e+00
#> 7.65 1.267651e+01 1.331152e+01 8.887237e+00 5.222936e+00 3.686946e+00
#> 8.8  6.389683e+00 8.887237e+00 1.331152e+01 1.192518e+01 1.039350e+01
#> 9.4  3.242977e+00 5.222936e+00 1.192518e+01 1.331152e+01 1.295051e+01
#> 9.7  2.127425e+00 3.686946e+00 1.039350e+01 1.295051e+01 1.331152e+01
#>