Compute a covariance matrix for multiple individuals

Compute the covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_cov(data, kern, hp, deriv = NULL)

Arguments

data

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), hp())
MagmaClustR:::list_kern_to_cov(db, "SE", hp)
#> $`1`
#>            1.1     2.55     3.05      3.1      3.9      6.6     7.25     7.45
#> 1.1  6.3131104 5.854795 5.508662 5.469800 4.766399 2.134577 1.627134 1.487659
#> 2.55 5.8547951 6.313110 6.256787 6.245023 5.913856 3.506627 2.859872 2.669661
#> 3.05 5.5086620 6.256787 6.313110 6.312545 6.151705 4.018351 3.354472 3.153895
#> 3.1  5.4698001 6.245023 6.312545 6.313110 6.169925 4.069449 3.405052 3.203747
#> 3.9  4.7663990 5.913856 6.151705 6.169925 6.313110 4.861304 4.222131 4.018351
#> 6.6  2.1345773 3.506627 4.018351 4.069449 4.861304 6.313110 6.218218 6.151705
#> 7.25 1.6271341 2.859872 3.354472 3.405052 4.222131 6.218218 6.313110 6.304065
#> 7.45 1.4876589 2.669661 3.153895 3.203747 4.018351 6.151705 6.304065 6.313110
#> 7.95 1.1742341 2.219632 2.669661 2.716724 3.506627 5.913856 6.203190 6.256787
#> 9.35 0.5503608 1.203316 1.521774 1.556392 2.176884 4.814049 5.389998 5.546806
#>          7.95      9.35
#> 1.1  1.174234 0.5503608
#> 2.55 2.219632 1.2033164
#> 3.05 2.669661 1.5217739
#> 3.1  2.716724 1.5563922
#> 3.9  3.506627 2.1768844
#> 6.6  5.913856 4.8140491
#> 7.25 6.203190 5.3899977
#> 7.45 6.256787 5.5468059
#> 7.95 6.313110 5.8847787
#> 9.35 5.884779 6.3131104
#> 
#> $`2`
#>            1.1     2.55     3.05      3.1      3.9      6.6     7.25     7.45
#> 1.1  6.3131104 5.854795 5.508662 5.469800 4.766399 2.134577 1.627134 1.487659
#> 2.55 5.8547951 6.313110 6.256787 6.245023 5.913856 3.506627 2.859872 2.669661
#> 3.05 5.5086620 6.256787 6.313110 6.312545 6.151705 4.018351 3.354472 3.153895
#> 3.1  5.4698001 6.245023 6.312545 6.313110 6.169925 4.069449 3.405052 3.203747
#> 3.9  4.7663990 5.913856 6.151705 6.169925 6.313110 4.861304 4.222131 4.018351
#> 6.6  2.1345773 3.506627 4.018351 4.069449 4.861304 6.313110 6.218218 6.151705
#> 7.25 1.6271341 2.859872 3.354472 3.405052 4.222131 6.218218 6.313110 6.304065
#> 7.45 1.4876589 2.669661 3.153895 3.203747 4.018351 6.151705 6.304065 6.313110
#> 7.95 1.1742341 2.219632 2.669661 2.716724 3.506627 5.913856 6.203190 6.256787
#> 9.35 0.5503608 1.203316 1.521774 1.556392 2.176884 4.814049 5.389998 5.546806
#>          7.95      9.35
#> 1.1  1.174234 0.5503608
#> 2.55 2.219632 1.2033164
#> 3.05 2.669661 1.5217739
#> 3.1  2.716724 1.5563922
#> 3.9  3.506627 2.1768844
#> 6.6  5.913856 4.8140491
#> 7.25 6.203190 5.3899977
#> 7.45 6.256787 5.5468059
#> 7.95 6.313110 5.8847787
#> 9.35 5.884779 6.3131104
#> 
#> $`3`
#>            1.1     2.55     3.05      3.1      3.9      6.6     7.25     7.45
#> 1.1  6.3131104 5.854795 5.508662 5.469800 4.766399 2.134577 1.627134 1.487659
#> 2.55 5.8547951 6.313110 6.256787 6.245023 5.913856 3.506627 2.859872 2.669661
#> 3.05 5.5086620 6.256787 6.313110 6.312545 6.151705 4.018351 3.354472 3.153895
#> 3.1  5.4698001 6.245023 6.312545 6.313110 6.169925 4.069449 3.405052 3.203747
#> 3.9  4.7663990 5.913856 6.151705 6.169925 6.313110 4.861304 4.222131 4.018351
#> 6.6  2.1345773 3.506627 4.018351 4.069449 4.861304 6.313110 6.218218 6.151705
#> 7.25 1.6271341 2.859872 3.354472 3.405052 4.222131 6.218218 6.313110 6.304065
#> 7.45 1.4876589 2.669661 3.153895 3.203747 4.018351 6.151705 6.304065 6.313110
#> 7.95 1.1742341 2.219632 2.669661 2.716724 3.506627 5.913856 6.203190 6.256787
#> 9.35 0.5503608 1.203316 1.521774 1.556392 2.176884 4.814049 5.389998 5.546806
#>          7.95      9.35
#> 1.1  1.174234 0.5503608
#> 2.55 2.219632 1.2033164
#> 3.05 2.669661 1.5217739
#> 3.1  2.716724 1.5563922
#> 3.9  3.506627 2.1768844
#> 6.6  5.913856 4.8140491
#> 7.25 6.203190 5.3899977
#> 7.45 6.256787 5.5468059
#> 7.95 6.313110 5.8847787
#> 9.35 5.884779 6.3131104
#>