Compute an inverse covariance matrix for multiple individuals

Compute the inverse covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_inv(db, kern, hp, pen_diag, deriv = NULL)

Arguments

db

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

pen_diag

A number. A jitter term, added on the diagonal to prevent numerical issues when inverting nearly singular matrices.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), MagmaClustR:::hp())
MagmaClustR:::list_kern_to_inv(db, "SE", hp, 0)
#> $`1`
#>              0.1          1.7          2.2         4.75         6.15
#> 0.1     1.556030    -6.255999     5.840676    -2.784955     6.275892
#> 1.7    -6.255999    33.853912   -33.518008    18.644119   -43.526578
#> 2.2     5.840676   -33.518008    33.947621   -20.149889    47.782720
#> 4.75   -2.784955    18.644119   -20.149889    17.787236   -47.941412
#> 6.15    6.275892   -43.526578    47.782720   -47.941412   142.915300
#> 7.65  -36.471502   258.748307  -286.945770   314.393061 -1017.336690
#> 8.25   88.485635  -631.828120   702.723324  -790.190802  2624.351185
#> 8.9  -211.433871  1518.564024 -1693.416351  1950.119695 -6638.012632
#> 9.1   171.832916 -1236.107597  1379.431815 -1599.010360  5480.597960
#> 9.65  -16.738602   120.889908  -135.149739   159.266070  -555.554765
#>             7.65          8.25          8.9          9.1        9.65
#> 0.1     -36.4715      88.48563    -211.4339     171.8329    -16.7386
#> 1.7     258.7483    -631.82812    1518.5640   -1236.1076    120.8899
#> 2.2    -286.9458     702.72332   -1693.4164    1379.4318   -135.1497
#> 4.75    314.3931    -790.19080    1950.1197   -1599.0104    159.2661
#> 6.15  -1017.3367    2624.35118   -6638.0126    5480.5980   -555.5548
#> 7.65   7981.4041  -21349.99180   56020.2929  -46750.3595   4873.6581
#> 8.25 -21349.9918   58020.43508 -154885.4202  129937.6286 -13740.3687
#> 8.9   56020.2929 -154885.42017  421753.4745 -356053.4419  38319.7808
#> 9.1  -46750.3595  129937.62863 -356053.4419  301205.0869 -32608.2481
#> 9.65   4873.6581  -13740.36866   38319.7808  -32608.2481   3592.0927
#> 
#> $`2`
#>              0.1          1.7          2.2         4.75         6.15
#> 0.1     1.556030    -6.255999     5.840676    -2.784955     6.275892
#> 1.7    -6.255999    33.853912   -33.518008    18.644119   -43.526578
#> 2.2     5.840676   -33.518008    33.947621   -20.149889    47.782720
#> 4.75   -2.784955    18.644119   -20.149889    17.787236   -47.941412
#> 6.15    6.275892   -43.526578    47.782720   -47.941412   142.915300
#> 7.65  -36.471502   258.748307  -286.945770   314.393061 -1017.336690
#> 8.25   88.485635  -631.828120   702.723324  -790.190802  2624.351185
#> 8.9  -211.433871  1518.564024 -1693.416351  1950.119695 -6638.012632
#> 9.1   171.832916 -1236.107597  1379.431815 -1599.010360  5480.597960
#> 9.65  -16.738602   120.889908  -135.149739   159.266070  -555.554765
#>             7.65          8.25          8.9          9.1        9.65
#> 0.1     -36.4715      88.48563    -211.4339     171.8329    -16.7386
#> 1.7     258.7483    -631.82812    1518.5640   -1236.1076    120.8899
#> 2.2    -286.9458     702.72332   -1693.4164    1379.4318   -135.1497
#> 4.75    314.3931    -790.19080    1950.1197   -1599.0104    159.2661
#> 6.15  -1017.3367    2624.35118   -6638.0126    5480.5980   -555.5548
#> 7.65   7981.4041  -21349.99180   56020.2929  -46750.3595   4873.6581
#> 8.25 -21349.9918   58020.43508 -154885.4202  129937.6286 -13740.3687
#> 8.9   56020.2929 -154885.42017  421753.4745 -356053.4419  38319.7808
#> 9.1  -46750.3595  129937.62863 -356053.4419  301205.0869 -32608.2481
#> 9.65   4873.6581  -13740.36866   38319.7808  -32608.2481   3592.0927
#> 
#> $`3`
#>              0.1          1.7          2.2         4.75         6.15
#> 0.1     1.556030    -6.255999     5.840676    -2.784955     6.275892
#> 1.7    -6.255999    33.853912   -33.518008    18.644119   -43.526578
#> 2.2     5.840676   -33.518008    33.947621   -20.149889    47.782720
#> 4.75   -2.784955    18.644119   -20.149889    17.787236   -47.941412
#> 6.15    6.275892   -43.526578    47.782720   -47.941412   142.915300
#> 7.65  -36.471502   258.748307  -286.945770   314.393061 -1017.336690
#> 8.25   88.485635  -631.828120   702.723324  -790.190802  2624.351185
#> 8.9  -211.433871  1518.564024 -1693.416351  1950.119695 -6638.012632
#> 9.1   171.832916 -1236.107597  1379.431815 -1599.010360  5480.597960
#> 9.65  -16.738602   120.889908  -135.149739   159.266070  -555.554765
#>             7.65          8.25          8.9          9.1        9.65
#> 0.1     -36.4715      88.48563    -211.4339     171.8329    -16.7386
#> 1.7     258.7483    -631.82812    1518.5640   -1236.1076    120.8899
#> 2.2    -286.9458     702.72332   -1693.4164    1379.4318   -135.1497
#> 4.75    314.3931    -790.19080    1950.1197   -1599.0104    159.2661
#> 6.15  -1017.3367    2624.35118   -6638.0126    5480.5980   -555.5548
#> 7.65   7981.4041  -21349.99180   56020.2929  -46750.3595   4873.6581
#> 8.25 -21349.9918   58020.43508 -154885.4202  129937.6286 -13740.3687
#> 8.9   56020.2929 -154885.42017  421753.4745 -356053.4419  38319.7808
#> 9.1  -46750.3595  129937.62863 -356053.4419  301205.0869 -32608.2481
#> 9.65   4873.6581  -13740.36866   38319.7808  -32608.2481   3592.0927
#>