Compute an inverse covariance matrix for multiple individuals

Compute the inverse covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_inv(db, kern, hp, pen_diag, deriv = NULL)

Arguments

db

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

pen_diag

A number. A jitter term, added on the diagonal to prevent numerical issues when inverting nearly singular matrices.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), MagmaClustR:::hp())
MagmaClustR:::list_kern_to_inv(db, "SE", hp, 0)
#> $`1`
#>             0.45          2.1         2.3          2.5        3.35         3.55
#> 0.45    3.564649    -387.2510    1027.970    -767.8065     527.188    -457.8943
#> 2.1  -387.250960   65546.5653 -180020.601  138732.3819 -106075.479   94029.1045
#> 2.3  1027.969549 -180020.6015  496505.222 -384187.1567  298281.390 -265233.9280
#> 2.5  -767.806470  138732.3819 -384187.157  298460.0088 -235322.709  209926.0730
#> 3.35  527.188041 -106075.4795  298281.390 -235322.7086  197910.349 -179125.6743
#> 3.55 -457.894344   94029.1045 -265233.928  209926.0730 -179125.674  162690.3379
#> 4.35   91.549889  -20101.5771   57313.289  -45875.2128   41281.221  -38005.8283
#> 4.8   -40.386589    9130.1523  -26159.253   21047.8882  -19437.070   18020.9634
#> 5.95    6.005592   -1432.7539    4143.494   -3367.5338    3275.855   -3082.4084
#> 6.5    -2.307743     560.4025   -1625.781    1325.8471   -1313.516    1242.5995
#>              4.35          4.8         5.95          6.5
#> 0.45     91.54989    -40.38659     6.005592    -2.307743
#> 2.1  -20101.57715   9130.15234 -1432.753937   560.402452
#> 2.3   57313.28933 -26159.25335  4143.493857 -1625.781264
#> 2.5  -45875.21276  21047.88816 -3367.533755  1325.847135
#> 3.35  41281.22061 -19437.06971  3275.855487 -1313.515760
#> 3.55 -38005.82831  18020.96345 -3082.408381  1242.599492
#> 4.35   9393.46367  -4596.71222   844.143905  -349.649244
#> 4.8   -4596.71222   2294.17572  -442.752602   187.193486
#> 5.95    844.14391   -442.75260   100.310273   -46.032158
#> 6.5    -349.64924    187.19349   -46.032158    22.550021
#> 
#> $`2`
#>             0.45          2.1         2.3          2.5        3.35         3.55
#> 0.45    3.564649    -387.2510    1027.970    -767.8065     527.188    -457.8943
#> 2.1  -387.250960   65546.5653 -180020.601  138732.3819 -106075.479   94029.1045
#> 2.3  1027.969549 -180020.6015  496505.222 -384187.1567  298281.390 -265233.9280
#> 2.5  -767.806470  138732.3819 -384187.157  298460.0088 -235322.709  209926.0730
#> 3.35  527.188041 -106075.4795  298281.390 -235322.7086  197910.349 -179125.6743
#> 3.55 -457.894344   94029.1045 -265233.928  209926.0730 -179125.674  162690.3379
#> 4.35   91.549889  -20101.5771   57313.289  -45875.2128   41281.221  -38005.8283
#> 4.8   -40.386589    9130.1523  -26159.253   21047.8882  -19437.070   18020.9634
#> 5.95    6.005592   -1432.7539    4143.494   -3367.5338    3275.855   -3082.4084
#> 6.5    -2.307743     560.4025   -1625.781    1325.8471   -1313.516    1242.5995
#>              4.35          4.8         5.95          6.5
#> 0.45     91.54989    -40.38659     6.005592    -2.307743
#> 2.1  -20101.57715   9130.15234 -1432.753937   560.402452
#> 2.3   57313.28933 -26159.25335  4143.493857 -1625.781264
#> 2.5  -45875.21276  21047.88816 -3367.533755  1325.847135
#> 3.35  41281.22061 -19437.06971  3275.855487 -1313.515760
#> 3.55 -38005.82831  18020.96345 -3082.408381  1242.599492
#> 4.35   9393.46367  -4596.71222   844.143905  -349.649244
#> 4.8   -4596.71222   2294.17572  -442.752602   187.193486
#> 5.95    844.14391   -442.75260   100.310273   -46.032158
#> 6.5    -349.64924    187.19349   -46.032158    22.550021
#> 
#> $`3`
#>             0.45          2.1         2.3          2.5        3.35         3.55
#> 0.45    3.564649    -387.2510    1027.970    -767.8065     527.188    -457.8943
#> 2.1  -387.250960   65546.5653 -180020.601  138732.3819 -106075.479   94029.1045
#> 2.3  1027.969549 -180020.6015  496505.222 -384187.1567  298281.390 -265233.9280
#> 2.5  -767.806470  138732.3819 -384187.157  298460.0088 -235322.709  209926.0730
#> 3.35  527.188041 -106075.4795  298281.390 -235322.7086  197910.349 -179125.6743
#> 3.55 -457.894344   94029.1045 -265233.928  209926.0730 -179125.674  162690.3379
#> 4.35   91.549889  -20101.5771   57313.289  -45875.2128   41281.221  -38005.8283
#> 4.8   -40.386589    9130.1523  -26159.253   21047.8882  -19437.070   18020.9634
#> 5.95    6.005592   -1432.7539    4143.494   -3367.5338    3275.855   -3082.4084
#> 6.5    -2.307743     560.4025   -1625.781    1325.8471   -1313.516    1242.5995
#>              4.35          4.8         5.95          6.5
#> 0.45     91.54989    -40.38659     6.005592    -2.307743
#> 2.1  -20101.57715   9130.15234 -1432.753937   560.402452
#> 2.3   57313.28933 -26159.25335  4143.493857 -1625.781264
#> 2.5  -45875.21276  21047.88816 -3367.533755  1325.847135
#> 3.35  41281.22061 -19437.06971  3275.855487 -1313.515760
#> 3.55 -38005.82831  18020.96345 -3082.408381  1242.599492
#> 4.35   9393.46367  -4596.71222   844.143905  -349.649244
#> 4.8   -4596.71222   2294.17572  -442.752602   187.193486
#> 5.95    844.14391   -442.75260   100.310273   -46.032158
#> 6.5    -349.64924    187.19349   -46.032158    22.550021
#>