Compute an inverse covariance matrix for multiple individuals

Compute the inverse covariance matrices associated with all individuals in the database, taking into account their specific inputs and hyper-parameters.

Usage

list_kern_to_inv(db, kern, hp, pen_diag, deriv = NULL)

Arguments

db

A tibble or data frame of input data. Required column: 'ID'. Suggested column: 'Input' (for indicating the reference input).

kern

A kernel function.

hp

A tibble or data frame, containing the hyper-parameters associated with each individual.

pen_diag

A number. A jitter term, added on the diagonal to prevent numerical issues when inverting nearly singular matrices.

deriv

A character, indicating according to which hyper-parameter the derivative should be computed. If NULL (default), the function simply returns the list of covariance matrices.

Value

A named list containing all of the inverse covariance matrices.

Examples

db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), MagmaClustR:::hp())
MagmaClustR:::list_kern_to_inv(db, "SE", hp, 0)
#> $`1`
#>                  0         1.4         2.05          2.5          4.1
#> 0        130475.47    -3734348     16298346    -18500490     79104643
#> 1.4    -3734347.69   109490546   -482810809    551794374  -2412649674
#> 2.05   16298345.60  -482810809   2138727386  -2451815553  10831562645
#> 2.5   -18500490.09   551794374  -2451815553   2816598443 -12532006508
#> 4.1    79104643.13 -2412649674  10831562645 -12532006508  57170257225
#> 4.35 -102588623.72  3139044665 -14114324301  16347612191 -74864159483
#> 5      46373952.91 -1430537075   6457261944  -7499403350  34685519713
#> 5.4   -17290674.59   535942164  -2424787154   2820739177 -13125022032
#> 7.5      263173.16    -8345386     38185664    -44781301    214830076
#> 8.45     -56380.12     1804228     -8293830      9759154    -47436258
#>              4.35            5          5.4          7.5         8.45
#> 0      -102588624     46373953    -17290675     263173.2    -56380.12
#> 1.4    3139044665  -1430537075    535942164   -8345386.3   1804227.90
#> 2.05 -14114324301   6457261944  -2424787154   38185663.7  -8293830.02
#> 2.5   16347612191  -7499403350   2820739177  -44781300.9   9759154.17
#> 4.1  -74864159483  34685519713 -13125022032  214830075.8 -47436258.41
#> 4.35  98093497615 -45519487121  17241269468 -283628811.8  62767384.13
#> 5    -45519487121  21209976840  -8054098048  134285280.7 -29897704.11
#> 5.4   17241269468  -8054098048   3063246542  -51509481.9  11513148.27
#> 7.5    -283628812    134285281    -51509482     908454.1   -207818.37
#> 8.45     62767384    -29897704     11513148    -207818.4     48124.43
#> 
#> $`2`
#>                  0         1.4         2.05          2.5          4.1
#> 0        130475.47    -3734348     16298346    -18500490     79104643
#> 1.4    -3734347.69   109490546   -482810809    551794374  -2412649674
#> 2.05   16298345.60  -482810809   2138727386  -2451815553  10831562645
#> 2.5   -18500490.09   551794374  -2451815553   2816598443 -12532006508
#> 4.1    79104643.13 -2412649674  10831562645 -12532006508  57170257225
#> 4.35 -102588623.72  3139044665 -14114324301  16347612191 -74864159483
#> 5      46373952.91 -1430537075   6457261944  -7499403350  34685519713
#> 5.4   -17290674.59   535942164  -2424787154   2820739177 -13125022032
#> 7.5      263173.16    -8345386     38185664    -44781301    214830076
#> 8.45     -56380.12     1804228     -8293830      9759154    -47436258
#>              4.35            5          5.4          7.5         8.45
#> 0      -102588624     46373953    -17290675     263173.2    -56380.12
#> 1.4    3139044665  -1430537075    535942164   -8345386.3   1804227.90
#> 2.05 -14114324301   6457261944  -2424787154   38185663.7  -8293830.02
#> 2.5   16347612191  -7499403350   2820739177  -44781300.9   9759154.17
#> 4.1  -74864159483  34685519713 -13125022032  214830075.8 -47436258.41
#> 4.35  98093497615 -45519487121  17241269468 -283628811.8  62767384.13
#> 5    -45519487121  21209976840  -8054098048  134285280.7 -29897704.11
#> 5.4   17241269468  -8054098048   3063246542  -51509481.9  11513148.27
#> 7.5    -283628812    134285281    -51509482     908454.1   -207818.37
#> 8.45     62767384    -29897704     11513148    -207818.4     48124.43
#> 
#> $`3`
#>                  0         1.4         2.05          2.5          4.1
#> 0        130475.47    -3734348     16298346    -18500490     79104643
#> 1.4    -3734347.69   109490546   -482810809    551794374  -2412649674
#> 2.05   16298345.60  -482810809   2138727386  -2451815553  10831562645
#> 2.5   -18500490.09   551794374  -2451815553   2816598443 -12532006508
#> 4.1    79104643.13 -2412649674  10831562645 -12532006508  57170257225
#> 4.35 -102588623.72  3139044665 -14114324301  16347612191 -74864159483
#> 5      46373952.91 -1430537075   6457261944  -7499403350  34685519713
#> 5.4   -17290674.59   535942164  -2424787154   2820739177 -13125022032
#> 7.5      263173.16    -8345386     38185664    -44781301    214830076
#> 8.45     -56380.12     1804228     -8293830      9759154    -47436258
#>              4.35            5          5.4          7.5         8.45
#> 0      -102588624     46373953    -17290675     263173.2    -56380.12
#> 1.4    3139044665  -1430537075    535942164   -8345386.3   1804227.90
#> 2.05 -14114324301   6457261944  -2424787154   38185663.7  -8293830.02
#> 2.5   16347612191  -7499403350   2820739177  -44781300.9   9759154.17
#> 4.1  -74864159483  34685519713 -13125022032  214830075.8 -47436258.41
#> 4.35  98093497615 -45519487121  17241269468 -283628811.8  62767384.13
#> 5    -45519487121  21209976840  -8054098048  134285280.7 -29897704.11
#> 5.4   17241269468  -8054098048   3063246542  -51509481.9  11513148.27
#> 7.5    -283628812    134285281    -51509482     908454.1   -207818.37
#> 8.45     62767384    -29897704     11513148    -207818.4     48124.43
#>